apps/backend/pyproject.toml
[project]
name = "bim-to-ai-backend"
version = "1.0.0"
description = "BIM-to-AI Pipeline Backend API"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Task Queue
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",

    # File Storage
    "minio>=7.2.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # Validation & Serialization
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",

    # Pipeline Dependencies (from existing)
    "ifcopenshell>=0.7.0",
    "pandas>=2.0.0",
    "networkx>=3.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pyarrow>=14.0.0",

    # Document Parsing
    "PyMuPDF>=1.23.0",          # PDF parsing
    "pdfplumber>=0.10.0",        # PDF tables
    "python-docx>=1.1.0",        # DOCX parsing
    "openpyxl>=3.1.0",           # XLSX parsing
    "python-pptx>=0.6.23",       # PPTX parsing
    "olefile>=0.46",             # HWP binary format
    "lxml>=5.1.0",               # XML parsing

    # bSDD API
    "httpx>=0.27.0",             # Async HTTP client

    # Utilities
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
